Create AssetImage feature to centralize picture management
