+++ inc/commondropdown.class.php
@@ -31,6 +31,7 @@
  */
 
 use Glpi\Application\View\TemplateRenderer;
+use Glpi\Features\AssetImage;
 
 if (!defined('GLPI_ROOT')) {
    die("Sorry. You can't access this file directly");
@@ -38,7 +39,7 @@ if (!defined('GLPI_ROOT')) {
 
 /// CommonDropdown class - generic dropdown
 abstract class CommonDropdown extends CommonDBTM {
-
+   use AssetImage;
 
    // From CommonDBTM
    public $dohistory                   = true;
@@ -921,86 +922,5 @@ abstract class CommonDropdown extends CommonDBTM {
 
    static function getIcon() {
       return "fas fa-edit";
-   }
-
-   /**
-    * Add/remove front and rear pictures for models
-    * @param  array $input the form input
-    * @return array        the altered input
-    */
-   function managePictures($input) {
-      foreach (['picture', 'picture_front', 'picture_rear'] as $name) {
-         if (isset($input["_blank_$name"])
-            && $input["_blank_$name"]) {
-            $input[$name] = '';
-
-            if (array_key_exists($name, $this->fields)) {
-               Toolbox::deletePicture($this->fields[$name]);
-            }
-         }
-
-         if (isset($input["_$name"])) {
-            $filename = array_shift($input["_$name"]);
-            $src      = GLPI_TMP_DIR . '/' . $filename;
-
-            $prefix   = null;
-            if (isset($input["_prefix_$name"])) {
-               $prefix = array_shift($input["_prefix_$name"]);
-            }
-
-            if ($dest = Toolbox::savePicture($src, $prefix)) {
-               $input[$name] = $dest;
-            } else {
-               Session::addMessageAfterRedirect(__('Unable to save picture file.'), true, ERROR);
-            }
-
-            if (array_key_exists($name, $this->fields)) {
-               Toolbox::deletePicture($this->fields[$name]);
-            }
-         }
-      }
-
-      $pictures = [];
-      if ($this->isField('pictures')) {
-         $input_keys = array_keys($input);
-         $pictures = importArrayFromDB($this->fields['pictures']);
-         $to_remove = [];
-         foreach ($input_keys as $input_key) {
-            if (strpos($input_key, '_blank_pictures_') === 0 && $input[$input_key]) {
-               $i = (int)str_replace('_blank_pictures_', '', $input_key);
-               if (isset($pictures[$i])) {
-                  Toolbox::deletePicture($pictures[$i]);
-                  $to_remove[] = $i;
-               }
-            }
-         }
-         $to_remove = array_reverse($to_remove);
-         foreach ($to_remove as $i) {
-            unset($pictures[$i]);
-         }
-      }
-
-      $new_pictures = [];
-      if (isset($input['_pictures'])) {
-         $pic_count = count($input['_pictures']);
-         if (!isset($input['pictures'])) {
-            $input['pictures'] = [];
-         }
-         for ($i = 0; $i < $pic_count; $i++) {
-            $filename = $input["_pictures"][$i];
-            $src = GLPI_TMP_DIR . '/' . $filename;
-
-            $prefix = $input["_prefix_pictures"][$i] ?? null;
-
-            if ($dest = Toolbox::savePicture($src, $prefix)) {
-               $new_pictures[$i] = $dest;
-            } else {
-               Session::addMessageAfterRedirect(__('Unable to save picture file.'), true, ERROR);
-            }
-         }
-      }
-      $input['pictures'] = exportArrayToDB(array_merge($pictures, $new_pictures));
-
-      return $input;
    }
 }
